+++ ExpandWorldData/ExpandWorldData.cs
@@ -11,7 +11,7 @@ public class EWD : BaseUnityPlugin
 {
   public const string GUID = "expand_world_data";
   public const string NAME = "Expand World Data";
-  public const string VERSION = "1.9";
+  public const string VERSION = "1.11";
 #nullable disable
   public static ManualLogSource Log;
   public static EWD Instance;
@@ -23,9 +23,9 @@ public class EWD : BaseUnityPlugin
     ModRequired = true,
     IsLocked = true
   };
-  public static string ConfigName = $"{GUID}.cfg";
-  public static bool NeedsMigration = File.Exists(Path.Combine(Paths.ConfigPath, "expand_world.cfg")) && !File.Exists(Path.Combine(Paths.ConfigPath, ConfigName));
+  public static bool NeedsMigration = File.Exists(Path.Combine(Paths.ConfigPath, "expand_world.cfg")) && !File.Exists(Path.Combine(Paths.ConfigPath, "expand_world_data.cfg"));
   public static string YamlDirectory = Path.Combine(Paths.ConfigPath, "expand_world");
+  public static string BackupDirectory = Path.Combine(Paths.ConfigPath, "expand_world_backups");
   public void InvokeRegenerate()
   {
     // Nothing to regenerate because the world hasn't been generated yet.
@@ -55,7 +55,7 @@ public class EWD : BaseUnityPlugin
     {
       if (Configuration.DataReload)
       {
-        SetupWatcher();
+        DataManager.SetupWatcher(Config);
         DataLoading.SetupWatcher();
         BiomeManager.SetupWatcher();
         LocationLoading.SetupWatcher();
@@ -83,7 +83,7 @@ public class EWD : BaseUnityPlugin
     Configuration.configLegacyGeneration.Value = true;
     Config.Save();
     var from = File.ReadAllLines(Path.Combine(Paths.ConfigPath, "expand_world.cfg"));
-    var to = File.ReadAllLines(Path.Combine(Paths.ConfigPath, ConfigName));
+    var to = File.ReadAllLines(Config.ConfigFilePath);
     foreach (var line in from)
     {
       var split = line.Split('=');
@@ -94,7 +94,7 @@ public class EWD : BaseUnityPlugin
           to[i] = line;
       }
     }
-    File.WriteAllLines(Path.Combine(Paths.ConfigPath, ConfigName), to);
+    File.WriteAllLines(Config.ConfigFilePath, to);
     Config.Reload();
   }
 #pragma warning disable IDE0051
@@ -104,22 +104,12 @@ public class EWD : BaseUnityPlugin
   }
 #pragma warning restore IDE0051
 
-  private void SetupWatcher()
-  {
-    FileSystemWatcher watcher = new(Paths.ConfigPath, ConfigName);
-    watcher.Changed += ReadConfigValues;
-    watcher.Created += ReadConfigValues;
-    watcher.Renamed += ReadConfigValues;
-    watcher.IncludeSubdirectories = true;
-    watcher.SynchronizingObject = ThreadingHelper.SynchronizingObject;
-    watcher.EnableRaisingEvents = true;
-  }
   private void YamlCleanUp()
   {
     try
     {
       if (!Directory.Exists(YamlDirectory)) return;
-      if (File.Exists(Path.Combine(Paths.ConfigPath, ConfigName))) return;
+      if (File.Exists(Config.ConfigFilePath)) return;
       Directory.Delete(YamlDirectory, true);
     }
     catch
@@ -127,21 +117,6 @@ public class EWD : BaseUnityPlugin
       Log.LogWarning("Failed to remove old yaml files.");
     }
   }
-  private void ReadConfigValues(object sender, FileSystemEventArgs e)
-  {
-    if (!File.Exists(Config.ConfigFilePath)) return;
-    try
-    {
-      Log.LogDebug("ReadConfigValues called");
-      Config.Reload();
-    }
-    catch
-    {
-      Log.LogError($"There was an issue loading your {Config.ConfigFilePath}");
-      Log.LogError("Please check your config entries for spelling and format!");
-    }
-  }
-
 }
 
 [HarmonyPatch(typeof(Terminal), nameof(Terminal.InitTerminal))]